feat: start quick setup modal integration
