+++ pages/community/setup-organization.tsx
@@ -11,10 +11,6 @@ import SetupOrganizationForm from "~community/common/components/organisms/Forms/
 import OnboardingLayout from "~community/common/components/templates/OnboardingLayout/OnboardingLayout";
 import ROUTES from "~community/common/constants/routes";
 import { FileTypes } from "~community/common/enums/CommonEnums";
-import {
-  MediaQueries,
-  useMediaQuery
-} from "~community/common/hooks/useMediaQuery";
 import { useTranslator } from "~community/common/hooks/useTranslator";
 import { themeSelector } from "~community/common/theme/themeSelector";
 import { ThemeTypes } from "~community/common/types/AvailableThemeColors";
@@ -24,8 +20,6 @@ import {
 } from "~community/common/types/CommonTypes";
 import { organizationSetupValidation } from "~community/common/utils/validation";
 import useGetCountryList from "~community/people/hooks/useGetCountryList";
-import { useCommonEnterpriseStore } from "~enterprise/common/store/commonStore";
-import { QuickSetupModalTypeEnums } from "~enterprise/quickSetup/enum/Common";
 
 const SetupOrganization: NextPage = () => {
   const router: NextRouter = useRouter();
@@ -34,17 +28,8 @@ const SetupOrganization: NextPage = () => {
 
   const translateText = useTranslator("onboarding", "organizationCreate");
 
-  const queryMatches = useMediaQuery();
-  const isBelow900 = queryMatches(MediaQueries.BELOW_900);
-
-  const { setQuickSetupModalType } = useCommonEnterpriseStore((state) => ({
-    setQuickSetupModalType: state.setQuickSetupModalType
-  }));
-
   const onSuccess = () => {
     router.replace(ROUTES.DASHBOARD.BASE);
-    !isBelow900 &&
-      setQuickSetupModalType(QuickSetupModalTypeEnums.START_QUICK_SETUP);
   };
 
   const { mutate: createOrganization, isPending } =
